tests: Add client script for the tests dispatcher service
After creating the parallel-tests-dispatcher service, it is needed
to modify the `run_test_concurrently.sh` script for using it. A
helper client allows the script to push the required tests to the
service and then iterate concurrently over that list, processing
the next available test until the list is depleted.

The change is backward compatible: the previous algorithm continues
working as it was, until two new arguments are added to the
`run_test_concurrently.sh` call, namely: BUILD_ID, and DISPATCHER_URL.
When these arguments are present, the dispatcher's client is used
instead of the previous partitioning algorithm.

It also includes a new LIZARDFS_TESTS_PATH argument that is expected
to point to the `lizardfs-tests` binary location.

+++ tests/tools/run_test_concurrently.sh
@@ -27,20 +27,22 @@
 
 set -o errexit -o nounset -o errtrace -o pipefail
 
-for ARGUMENT in "$@"
-do
+for ARGUMENT in "$@"; do
 	KEY=$(echo $ARGUMENT | cut -f1 -d=)
 	VALUE=$(echo $ARGUMENT | cut -f2 -d=)
 
 	case "$KEY" in
-		WORKSPACE)      WORKSPACE=${VALUE} ;;
-		TEST_SUITE)     TEST_SUITE=${VALUE} ;;
-		EXCLUDE_TESTS)  EXCLUDE_TESTS=${VALUE} ;;
-		RUN_UNITTESTS)  RUN_UNITTESTS=${VALUE} ;;
-		NODES_COUNT)    NODES_COUNT=${VALUE} ;;
-		NODE_NUMBER)    NODE_NUMBER=${VALUE} ;;
-		VALGRIND)       VALGRIND=${VALUE} ;;
-		*)
+		WORKSPACE) WORKSPACE=${VALUE} ;;
+		TEST_SUITE) TEST_SUITE=${VALUE} ;;
+		EXCLUDE_TESTS) EXCLUDE_TESTS=${VALUE} ;;
+		RUN_UNITTESTS) RUN_UNITTESTS=${VALUE} ;;
+		NODES_COUNT) NODES_COUNT=${VALUE} ;;
+		NODE_NUMBER) NODE_NUMBER=${VALUE} ;;
+		VALGRIND) VALGRIND=${VALUE} ;;
+		BUILD_ID) BUILD_ID=${VALUE} ;;
+		DISPATCHER_URL) DISPATCHER_URL=${VALUE} ;;
+		LIZARDFS_TESTS_PATH) LIZARDFS_TESTS_PATH=${VALUE} ;;
+		*) ;;
 	esac
 done
 
@@ -50,13 +52,20 @@ NODES_COUNT=${NODES_COUNT:-1}
 NODE_NUMBER=${NODE_NUMBER:-1}
 RUN_UNITTESTS=${RUN_UNITTESTS:-'false'}
 VALGRIND=${VALGRIND:-'No'}
+BUILD_ID=${BUILD_ID:-}
+DISPATCHER_URL=${DISPATCHER_URL:-}
+LIZARDFS_TESTS_PATH=${LIZARDFS_TESTS_PATH:-"${WORKSPACE}/install/lizardfs/bin/lizardfs-tests"}
 
 export LIZARDFS_ROOT=$WORKSPACE/install/lizardfs
 export TEST_OUTPUT_DIR=$WORKSPACE/test_output
 export TERM=xterm
 
+FG_GREEN=$(tput setaf 2)
+FG_RED=$(tput setaf 1)
+RESET=$(tput sgr0)
+
 if [ $VALGRIND == "Yes" ]; then
-	export USE_VALGRIND=YES;
+	export USE_VALGRIND=YES
 fi
 
 # TODO: maybe somehow implement the following lines, to not run compilation,
@@ -71,14 +80,87 @@ make -C build/lizardfs -j$(nproc) install
 
 killall -9 lizardfs-tests || true
 mkdir -m 777 -p $TEST_OUTPUT_DIR
-rm -rf $TEST_OUTPUT_DIR/* || true
+rm -rf "${TEST_OUTPUT_DIR:?}"/* || true
 rm -rf /mnt/ramdisk/* || true
 
-FILTER=$(python3 $WORKSPACE/tests/tools/filter_tests.py --workspace $WORKSPACE --test_suite $TEST_SUITE \
-	--excluded_tests $EXCLUDE_TESTS --nodes_count $NODES_COUNT --node_number $NODE_NUMBER)
+if [ -z "${DISPATCHER_URL}" ]; then
 
-if [ $RUN_UNITTESTS == "true" ]; then
-	$WORKSPACE/build/lizardfs/src/unittests/unittests --gtest_color=yes --gtest_output=xml:$TEST_OUTPUT_DIR/unit_test_results.xml
+	FILTER=$(python3 $WORKSPACE/tests/tools/filter_tests.py --workspace $WORKSPACE --test_suite $TEST_SUITE \
+		--excluded_tests $EXCLUDE_TESTS --nodes_count $NODES_COUNT --node_number $NODE_NUMBER)
+
+	if [ $RUN_UNITTESTS == "true" ]; then
+		$WORKSPACE/build/lizardfs/src/unittests/unittests --gtest_color=yes --gtest_output=xml:$TEST_OUTPUT_DIR/unit_test_results.xml
+	fi
+
+	$LIZARDFS_ROOT/bin/lizardfs-tests --gtest_color=yes --gtest_filter=$FILTER --gtest_output=xml:$TEST_OUTPUT_DIR/test_results.xml
+
+else
+	{
+		export TESTS_DISPATCHER_URL="${DISPATCHER_URL}"
+		python3 "${WORKSPACE}/tests/dispatcher/client/dispatcher_client.py" \
+			--action push_list \
+			--build_id "${BUILD_ID}" \
+			--lizardfs_tests_path "${LIZARDFS_TESTS_PATH}" \
+			--test_suite "${TEST_SUITE}" \
+			--excluded_tests "${EXCLUDE_TESTS}"
+
+		get_next_test() {
+			echo -n "$(python3 "${WORKSPACE}/tests/dispatcher/client/dispatcher_client.py" \
+				--action next_test \
+				--build_id "${BUILD_ID}" \
+				--test_suite "${TEST_SUITE}")"
+		}
+
+		conditionally_pluralize() {
+			[ "${1:-1}" == "1" ] && echo "" || echo "s"
+		}
+
+		is_set() {
+			option=${1:-}
+			set -o | grep "${option}" | grep -q "on"
+		}
+
+		declare -a failed_tests=()
+		tests_count=0
+		NEXT_TEST=$(get_next_test)
+		START_TIME=$(($(date +%s%N) / 1000000))
+		while [ "${NEXT_TEST}" != "" ]; do
+			"${LIZARDFS_ROOT}/bin/lizardfs-tests" \
+				--gtest_color=yes \
+				--gtest_filter="${TEST_SUITE}.${NEXT_TEST}" \
+				--gtest_output="xml:${TEST_OUTPUT_DIR}/test_results.xml" || {
+				failed_tests+=("${TEST_SUITE}.${NEXT_TEST}")
+				true
+			}
+			NEXT_TEST=$(get_next_test)
+			((++tests_count))
+		done
+
+		xtrace_was_enabled=false
+		if is_set xtrace; then
+			xtrace_was_enabled=true
+			set +x
+		fi
+		END_TIME=$(($(date +%s%N) / 1000000))
+		ELAPSED_TIME=$((END_TIME - START_TIME))
+		failed_tests_count=${#failed_tests[@]}
+		passed_tests_count=$((tests_count - failed_tests_count))
+
+		echo "${FG_GREEN}[----------]${RESET} Global test environment tear-down"
+		echo "${FG_GREEN}[==========]${RESET} ${tests_count} test$(conditionally_pluralize "${tests_count}") from 1 test case ran. (${ELAPSED_TIME} ms total)"
+		echo "${FG_GREEN}[  PASSED  ]${RESET} ${passed_tests_count} test$(conditionally_pluralize "${passed_tests_count}")."
+		if ((failed_tests_count > 0)); then
+			echo "${FG_RED}[  FAILED  ]${RESET} ${failed_tests_count} test$(conditionally_pluralize "${failed_tests_count}"), listed below:"
+			for failed_test in "${failed_tests[@]}"; do
+				echo "${FG_RED}[  FAILED  ]${RESET} ${failed_test}"
+			done
+		fi
+		echo ""
+		echo " ${failed_tests_count} FAILED TEST"
+		echo "Build step 'Execute shell' marked build as failure"
+		if ${xtrace_was_enabled}; then
+			set -x
+		fi
+		exit "${failed_tests_count}"
+	}
 fi
-
-$LIZARDFS_ROOT/bin/lizardfs-tests --gtest_color=yes --gtest_filter=$FILTER --gtest_output=xml:$TEST_OUTPUT_DIR/test_results.xml
